more tests

+++ dot-net-sdk/dto/Allocation.cs
@@ -1,5 +1,5 @@
 namespace eppo_sdk.dto;
 
-public record Allocation(string key, List<Rule> rules, List<Split> splits, bool doLog, int? startAt, int? endAt)
+public record Allocation(string key, List<Rule> rules, List<Split> splits, bool doLog, long? startAt, long? endAt)
 {
 }
